Refactor: Move vnic profile pool from oVirt DB to eduVirt DB
