+++ src/main/java/pl/lodz/p/it/eduvirt/entity/eduvirt/network/VnicProfilePool.java
@@ -1,12 +1,8 @@
-package pl.lodz.p.it.eduvirt.entity.ovirt;
+package pl.lodz.p.it.eduvirt.entity.eduvirt.network;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.PrimaryKeyJoinColumn;
 import jakarta.persistence.Table;
 import jakarta.persistence.Version;
 import lombok.EqualsAndHashCode;
@@ -14,12 +10,11 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import pl.lodz.p.it.eduvirt.entity.ovirt.readonly.VnicProfile;
 
 import java.util.UUID;
 
 @Entity
-@Table(name = "i72_vnic_profile_pool", schema = "eduvirt")
+@Table(name = "i72_vnic_profile_pool")
 @Getter
 @ToString
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
@@ -32,14 +27,6 @@ public class VnicProfilePool {
     @Id
     @Column(name = "id", unique = true, nullable = false, updatable = false)
     private UUID id;
-
-    @OneToOne(optional = false, fetch = FetchType.EAGER, cascade = CascadeType.REFRESH)
-    @PrimaryKeyJoinColumn(
-            name = "id",
-            referencedColumnName = "id"
-//            foreignKey = @ForeignKey(name = "vnic_profile_pool_id_fk")
-    )
-    private VnicProfile vnicProfile;
 
     @Version
     @Column(name = "version", nullable = false)
