+++ src/main/java/pl/lodz/p/it/eduvirt/repository/eduvirt/VnicProfileRepository.java
@@ -1,8 +1,8 @@
-package pl.lodz.p.it.eduvirt.repository.ovirt;
+package pl.lodz.p.it.eduvirt.repository.eduvirt;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pl.lodz.p.it.eduvirt.entity.ovirt.VnicProfilePool;
+import pl.lodz.p.it.eduvirt.entity.eduvirt.network.VnicProfilePool;
 
 import java.util.UUID;
